Move Variant types to Apache.Arrow.Scalars.Variant sub-namespace
Move all Variant*.cs files into a Variant/ subdirectory within
src/Apache.Arrow.Scalars and update the namespace from
Apache.Arrow.Scalars to Apache.Arrow.Scalars.Variant.

Update using directives in all consumers: Operations, Tests, and
Benchmarks projects.
