+++ src/Apache.Arrow.Scalars/Variant/VariantBuilder.cs
@@ -16,7 +16,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Apache.Arrow.Scalars
+namespace Apache.Arrow.Scalars.Variant
 {
     /// <summary>
     /// Encodes a <see cref="VariantValue"/> into the binary variant format,
