+++ src/Apache.Arrow.Scalars/Variant/VariantMetadataBuilder.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Apache.Arrow.Scalars
+namespace Apache.Arrow.Scalars.Variant
 {
     /// <summary>
     /// Collects unique field names and builds the sorted binary metadata
