+++ src/Apache.Arrow.Scalars/Variant/VariantReader.cs
@@ -20,7 +20,7 @@ using System.Data.SqlTypes;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace Apache.Arrow.Scalars
+namespace Apache.Arrow.Scalars.Variant
 {
     /// <summary>
     /// Zero-copy reader for a single variant value. Provides type inspection
