+++ src/Apache.Arrow.Scalars/Variant/VariantValue.cs
@@ -20,7 +20,7 @@ using System.Globalization;
 using System.Linq;
 using System.Runtime.CompilerServices;
 
-namespace Apache.Arrow.Scalars
+namespace Apache.Arrow.Scalars.Variant
 {
     /// <summary>
     /// Represents a materialized variant value as a discriminated union.
