+++ test/Apache.Arrow.Scalars.Tests/VariantBuilderTests.cs
@@ -17,6 +17,8 @@ using System;
 using System.Collections.Generic;
 using Xunit;
 
+using Apache.Arrow.Scalars.Variant;
+
 namespace Apache.Arrow.Scalars.Tests
 {
     public class VariantBuilderTests
