+++ test/Apache.Arrow.Scalars.Tests/VariantReaderObjectTests.cs
@@ -17,6 +17,8 @@ using System;
 using System.Text;
 using Xunit;
 
+using Apache.Arrow.Scalars.Variant;
+
 namespace Apache.Arrow.Scalars.Tests
 {
     public class VariantReaderObjectTests
