+++ test/Apache.Arrow.Variant.Benchmarks/CreationBenchmarks.cs
@@ -15,6 +15,8 @@
 
 using BenchmarkDotNet.Attributes;
 
+using Apache.Arrow.Scalars.Variant;
+
 namespace Apache.Arrow.Scalars.Benchmarks
 {
     /// <summary>
