+++ test/Apache.Arrow.Variant.Benchmarks/MixedWorkloadBenchmarks.cs
@@ -16,6 +16,8 @@
 using System.Collections.Generic;
 using BenchmarkDotNet.Attributes;
 
+using Apache.Arrow.Scalars.Variant;
+
 namespace Apache.Arrow.Scalars.Benchmarks
 {
     /// <summary>
